extract deprecations and fonts from orchestration
