+++ src/js/themes/hpe.js
@@ -57,6 +57,8 @@ import { buildMiscTheme } from './misc';
 import { buildContentTheme } from './content';
 import { buildFormTheme } from './form';
 import { buildLayoutTheme } from './layout';
+import { buildDeprecations } from './deprecations';
+import { getGraphikFamily, getGraphikFontFaces } from './fonts';
 
 const buildTheme = (tokens, flags) => {
   const { light, dark, large, global, components } = tokens;
@@ -138,109 +140,8 @@ const buildTheme = (tokens, flags) => {
     },
   });
   const layoutTheme = buildLayoutTheme(tokens);
-
-  const referenceColorDocumentationMessage =
-    'Please reference the HPE Design System color documentation (https://design-system.hpe.design/foundation/color) for a list of available colors';
-
-  const backgroundDeprecationMessage = (background) =>
-    `The background '${background}' is deprecated and will be removed in v10 to ensure access to the latest Brand assets. Please replace this key by referencing an image URL directly. You can find approved backgrounds within HPE Brand Central (https://brandcentral.hpe.com/brand-central/content/imagery).`;
-
-  const colorRemovedDeprecationMessage = (color, alt) =>
-    `The color '${color}' is deprecated and will be removed in v10. ${alt ? `Please use '${alt}' instead` : referenceColorDocumentationMessage}.`;
-
-  const colorUndefinedDeprecationMessage = (color, future, alt) =>
-    `The color '${color}' is deprecated and ${future ? "will be set to 'undefined' in v10" : "its value is set to 'undefined'"}. ${alt ? `Please use '${alt}' instead` : referenceColorDocumentationMessage}.`;
-
-  const deprecationEntry = (name, message) => ({ name, message });
-
-  // Deprecated backgrounds
-  const deprecatedBackgrounds = [
-    'datawave-green-1',
-    'datawave-green-2',
-    'datawave-multi-1',
-    'datawave-multi-2',
-    'datawave-multi-3',
-    'datawave-multi-4',
-    'datawave-multi-5',
-    'datawave-multi-6',
-    'datawave-white-1',
-    'datawave-white-2',
-    'datawave-white-3',
-    'datawave-white-4',
-    'light-shadow-1',
-    'light-shadow-2',
-    'light-shadow-3',
-    'light-shadow-4',
-    'orange-yellow',
-    'purple-blue',
-    'purple-blue-yellow',
-    'purple-magenta-yellow',
-  ].map((name) => deprecationEntry(name, backgroundDeprecationMessage(name)));
-
-  // Deprecated button kinds
-  const deprecatedButtonKinds = [
-    { name: 'cta-primary', alt: 'primary' },
-    { name: 'cta-alternate', alt: 'secondary' },
-  ].map(({ name, alt }) =>
-    deprecationEntry(
-      name,
-      `The "${name}" button kind is deprecated and will be removed in v10. Please use "${alt}" instead.`,
-    ),
-  );
-
-  // Deprecated colors
-  const deprecatedColors = [
-    // Undefined without alternative
-    ...[
-      'accent-1',
-      'accent-2',
-      'accent-3',
-      'accent-4',
-      'neutral-1',
-      'neutral-2',
-      'neutral-3',
-      'neutral-4',
-      'neutral-5',
-      'status-error',
-    ].map((name) =>
-      deprecationEntry(name, colorUndefinedDeprecationMessage(name)),
-    ),
-    // Undefined (future) with alternative - graph colors
-    ...[0, 1, 2, 3, 4, 5, 6, 7].map((i) =>
-      deprecationEntry(
-        `graph-${i}`,
-        colorUndefinedDeprecationMessage(
-          `graph-${i}`,
-          true,
-          `dataVis-categorical-${(i + 1) * 10}`,
-        ),
-      ),
-    ),
-    // Undefined (future) with alternative - other
-    deprecationEntry(
-      'status-disabled',
-      colorUndefinedDeprecationMessage('status-disabled', true, 'text-weak'),
-    ),
-    // Removed with alternative
-    deprecationEntry(
-      'disabled-text',
-      colorRemovedDeprecationMessage('disabled-text', 'text-disabled'),
-    ),
-    // Removed without alternative
-    deprecationEntry(
-      'background-cta-alternate',
-      colorRemovedDeprecationMessage('background-cta-alternate'),
-    ),
-  ];
-
-  // Figure out what name is being used for the Graphik font in the family list.
-  // Since we're going to map the font-faces to Graphik font files we specifically
-  // look for something containing "Graphik"
-  const family =
-    global.hpe.fontStack.primary
-      .split(',')
-      .map((s) => s.trim().replace(/['"]/g, ''))
-      .find((f) => f.includes('Graphik')) || 'HPE Graphik';
+  const deprecations = buildDeprecations();
+  const family = getGraphikFamily(global.hpe.fontStack.primary);
 
   return deepFreeze({
     defaultMode: 'light',
@@ -259,11 +160,11 @@ const buildTheme = (tokens, flags) => {
         disabled: { opacity: 0.3 },
       },
       deprecated: {
-        backgrounds: deprecatedBackgrounds,
+        backgrounds: deprecations.backgrounds,
         button: {
-          kind: deprecatedButtonKinds,
+          kind: deprecations.buttonKinds,
         },
-        colors: deprecatedColors,
+        colors: deprecations.colors,
       },
       input: {
         font: {
@@ -312,65 +213,7 @@ const buildTheme = (tokens, flags) => {
       },
       font: {
         family: global.hpe.fontStack.primary,
-        face: `
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Regular-Web.woff2") format('woff2');
-          }
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Regular-Web.woff2") format('woff2');
-            font-weight: 400;
-          }
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Bold-Web.woff2") format('woff2');
-            font-weight: 700;
-          }
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Semibold-Web.woff2") format('woff2');
-            font-weight: 600;
-          }
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Medium-Web.woff2") format('woff2');
-            font-weight: 500;
-          }
-          @font-face {
-            font-family: "${family}";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphik-Extralight-Web.woff2") format('woff2');
-            font-weight: 100;
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Regular-Web.woff2") format('woff2');
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Regular-Web.woff2") format('woff2');
-            font-weight: 400;
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Bold-Web.woff2") format('woff2');
-            font-weight: 700;
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Semibold-Web.woff2") format('woff2');
-            font-weight: 600;
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Medium-Web.woff2") format('woff2');
-            font-weight: 500;
-          }
-          @font-face {
-            font-family: "GraphikXXCondensed";
-            src: url("https://www.hpe.com/content/dam/hpe/fonts/graphik/HPEGraphikXXCondensed-Extralight-Web.woff2") format('woff2');
-            font-weight: 100;
-          }`,
+        face: getGraphikFontFaces(family),
         size: large.hpe.text.medium.fontSize,
         height: large.hpe.text.medium.lineHeight,
       },
